+++ src/main/java/com/umc/owncast/domain/playlist/controller/PlaylistController.java
@@ -2,10 +2,7 @@ package com.umc.owncast.domain.playlist.controller;
 
 import com.umc.owncast.common.response.ApiResponse;
 import com.umc.owncast.domain.playlist.dto.PlaylistDTO;
-import com.umc.owncast.domain.playlist.service.PlaylistAddServiceImpl;
-import com.umc.owncast.domain.playlist.service.PlaylistDeleteServiceImpl;
-import com.umc.owncast.domain.playlist.service.PlaylistGetServiceImpl;
-import com.umc.owncast.domain.playlist.service.PlaylistModifyServiceImpl;
+import com.umc.owncast.domain.playlist.service.*;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,16 +14,13 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class PlaylistController {
 
-    private final PlaylistAddServiceImpl PlaylistAddService;
-    private final PlaylistModifyServiceImpl PlaylistModifyService;
-    private final PlaylistDeleteServiceImpl PlaylistDeleteService;
-    private final PlaylistGetServiceImpl PlaylistGetService;
+    private final PlaylistService PlaylistService;
 
     @CrossOrigin
     @Operation(summary = "카테고리 추가")
     @PostMapping("/playlist")
     public ApiResponse<PlaylistDTO.AddPlaylistDTO> addPlaylist(@RequestParam("playlistName") String playlistName) {
-        return ApiResponse.onSuccess(PlaylistAddService.addPlaylist(playlistName));
+        return ApiResponse.onSuccess(PlaylistService.addPlaylist(playlistName));
     }
 
     @CrossOrigin
@@ -34,14 +28,14 @@ public class PlaylistController {
     @PatchMapping("/playlist/{playlistId}")
     public ApiResponse<PlaylistDTO.ModifyPlaylistDTO> modifyPlaylist(@RequestParam("playlistId") Long playlistId,
                                                                      @RequestParam("playlistName") String playlistName) {
-        return ApiResponse.onSuccess(PlaylistModifyService.modifyPlaylist(playlistId, playlistName));
+        return ApiResponse.onSuccess(PlaylistService.modifyPlaylist(playlistId, playlistName));
     }
 
     @CrossOrigin
     @Operation(summary = "카테고리 삭제")
     @DeleteMapping("/playlist/{playlistId}")
     public ApiResponse<PlaylistDTO.DeletePlaylistDTO> deletePlaylist(@RequestParam("playlistId") Long playlistId) {
-        return ApiResponse.onSuccess(PlaylistDeleteService.deletePlaylist(playlistId));
+        return ApiResponse.onSuccess(PlaylistService.deletePlaylist(playlistId));
     }
 
     @CrossOrigin
@@ -50,6 +44,6 @@ public class PlaylistController {
     public ApiResponse<PlaylistDTO.GetPlaylistDTO> getPlaylist(@RequestParam("playlistId") Long playlistId,
                                                                @RequestParam("page") Integer page,
                                                                @RequestParam("size") Integer size) {
-        return ApiResponse.onSuccess(PlaylistGetService.getPlaylist(playlistId, page, size));
+        return ApiResponse.onSuccess(PlaylistService.getPlaylist(playlistId, page, size));
     }
 }
