--- src/main/java/com/umc/owncast/domain/playlist/service/PlaylistDeleteService.java
@@ -1,4 +1,0 @@
-package com.umc.owncast.domain.playlist.service;
-
-public interface PlaylistDeleteService {
-}
