--- src/main/java/com/umc/owncast/domain/playlist/service/PlaylistModifyService.java
@@ -1,4 +1,0 @@
-package com.umc.owncast.domain.playlist.service;
-
-public interface PlaylistModifyService {
-}
